refactor: make some renaming according pull request
