+++ src/App.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import styles from './App.module.scss';
 import './styles/fonts.scss';
 import Route from './utils/Route';
-import HomePage from './components/Pages/Home';
-import CheckoutPage from './components/Pages/Checkout';
-import ProductPage from './components/Pages/Product';
+import HomePage from './components/Pages/HomePage';
+import CheckoutPage from './components/Pages/CheckoutPage';
+import ProductPage from './components/Pages/ProductPage';
 
 const App = () => {
 
@@ -19,6 +19,7 @@ const App = () => {
       <Route path="/product">
         <ProductPage />
       </Route>
+
     </div>
   );
 };
